requirements.txt
psycopg2-binary
pytest==8.3.2